Changed float to int in health, using damage note

+++ Solution/Entity/HealthComponentData.h
@@ -3,5 +3,5 @@
 struct HealthComponentData
 {
 	bool myExistsInEntity = false;
-	float myMaxHealth;
+	int myMaxHealth;
 };
